next-audit-python/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==4.0.0
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
entrypoints==0.4
faiss-cpu==1.7.4
flask==3.0.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
langchain==0.0.184
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.1
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
protobuf==3.20.3
pyarrow==14.0.0
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
pypdf2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
semver==3.0.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
streamlit==1.18.1
sympy==1.12
tenacity==8.2.3
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
